requirements.txt
phasellm
python-dotenv
Django
